Split withdraw tokens and items
