+++ src/features/game/actions/withdraw.ts
@@ -42,6 +42,10 @@ export async function withdraw({
     throw new Error(ERRORS.TOO_MANY_REQUESTS);
   }
 
+  if (response.status >= 400) {
+    throw new Error(ERRORS.FAILED_REQUEST);
+  }
+
   const transaction = await response.json();
 
   const newSessionId = await metamask.getSessionManager().withdraw(transaction);
